+++ src/assets/gql/seventv.user.gql.ts
@@ -19,6 +19,30 @@ export const actorQuery = gql`
 	}
 `;
 
+export const userQuery = gql`
+	query GetUser($id: ObjectID!) {
+		user: user(id: $id) {
+			id
+			username
+			display_name
+			avatar_url
+			style {
+				color
+				paint_id
+			}
+		}
+	}
+`;
+
+export namespace userQuery {
+	export interface Result {
+		user: SevenTV.User;
+	}
+	export interface Variables {
+		id: string;
+	}
+}
+
 export namespace actorQuery {
 	export interface Result {
 		user: SevenTV.User;